moving old formatting to qsub module
